Remove `encode` module
Superseded by objc2-encode
